+++ src/models.rs
@@ -5,16 +5,18 @@ use diesel::prelude::*;
 #[diesel(check_for_backend(diesel::sqlite::Sqlite))]
 pub struct Calculation {
     pub id: String,
-    pub time: i32,
-    pub distance: i32,
-    pub pace: i32,
+    // if more precision is needed, store them as text because Diesel does not support
+    // f64 due to potential precision issues. I doubt more precision would be needed tho.
+    pub time: f32,
+    pub distance: f32,
+    pub pace: f32,
 }
 
 #[derive(Insertable)]
 #[diesel(table_name = crate::schema::calculations)]
 pub struct NewCalculation<'a> {
     pub id: &'a str,
-    pub time: i32,
-    pub distance: i32,
-    pub pace: i32,
+    pub time: f32,
+    pub distance: f32,
+    pub pace: f32,
 }
